update and restruct the model fill
