+++ main/Main.py
@@ -1,20 +1,24 @@
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-
-
+from model.Modles import db
+from model.Modles import Store
 # TODO: make the app config to connect it to database
 app = Flask(__name__)
 
-# adding configuration for using a sqlite database
-app.config['SQLALCHEMY_DATABASE_URI'] ="mysql+pymysql://root:@localhost:3306/product"
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-db = SQLAlchemy(app)
-db.create_all()
+
+# adding configuration for using a sqlite database
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:@localhost/product'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+db.init_app(app)
+
 
 @app.route('/')
 def index():
-    return ""
+    try:
+        s = Store.query.filter_by(name='ss').first()
+        return s.name
+    except Exception as e:
+        return f"An error occurred: {str(e)}"
 
 
 if __name__ == '__main__':
